make window size and stride parameters
Make the window size and stride over the audio file parameters

Note: This does not change the window size and stride per spectrogram.

This, instead, modifies how big the examples of the file are.

This also modifies how they overlap.

The spectrograms consider each window as its own data.

It then puts its own window size and stride over it.